using double

+++ Laboratory_I/ex2/src/lab1_ex2_lib.c
@@ -6,14 +6,14 @@
 double mean(dtype *v, int n){
 	double sum = 0;
 	for (int i = 0; i < n; i++)
-		sum += (double) v[i];
-	return sum / (double)n;
+		sum += v[i];
+	return sum / n;
 }
 
 double variance(dtype *v, double mean, int n){
 	double sum = 0, mu_diff;
 	for (int i = 0; i < n; i++){
-		mu_diff = (((double)v[i]) - mean);
+		mu_diff = (v[i] - mean);
 		sum += mu_diff*mu_diff;
 	}
 	return sum/n;
